+++ src/services/path.js
@@ -1,6 +1,12 @@
-'use strict';
-
-exports.generate = function (path, opts) {
-  var pathPrefix = opts.expressParentApp ? '/' : '/forest/';
+exports.generate = (path, options) => {
+  var pathPrefix = options.expressParentApp ? '/' : '/forest/';
   return pathPrefix + path;
+};
+
+exports.generateForSmartActionCustomEndpoint = (path, options) => {
+  if (options.expressParentApp) {
+    return path.replace(/^\/forest/, '');
+  }
+
+  return path;
 };